mental_health_backend/.history/mental_health_backend/requirements_20250414135103.txt
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
flask-migrate==4.0.4
flask-pymongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
openai==0.27.8
requests==2.31.0
gunicorn==21.2.0